+++ docs/CHANGELOG.md
@@ -3,4 +3,6 @@ Changelog
 
 1. Initialize repository with project 1 05-04-2019.
 
-2. Second Project added 01-06-2019. 
+2. Second Project Countdown timer added 01-06-2019. 
+
+3. Third project battery tester added 15-06-2019